Fixed passport-beam from comments
