+++ passport-beam/passport-beam-tests.ts
@@ -1,4 +1,4 @@
-/// <reference path="./passport-beam.d.ts" />
+/// <reference path="./index.d.ts" />
 
 /**
  * Created by AtlasDev on 4/10/2016.
